% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerscale.R
\name{as.data.frame.powerscaled_draws_summary}
\alias{as.data.frame.powerscaled_draws_summary}
\title{Convert power-scaled draws to data.frame}
\usage{
\method{as.data.frame}{powerscaled_draws_summary}(x, ...)
}
\arguments{
\item{x}{powerscaled_draws}

\item{...}{unused}
}
\description{
Convert power-scaled draws to data.frame
}
