% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_log_prior.R
\name{calculate_log_prior}
\alias{calculate_log_prior}
\alias{calculate_log_prior.stanfit}
\alias{calculate_log_prior.brmsfit}
\title{Log prior calculation}
\usage{
calculate_log_prior(x, ...)

\method{calculate_log_prior}{stanfit}(x, ...)

\method{calculate_log_prior}{brmsfit}(x, ...)
}
\arguments{
\item{x}{Fitted model object.}

\item{...}{Currently unused.}
}
\value{
A draws_array object containing calculated log_prior values.
}
\description{
Log prior calculation by subtracting log likelihood from log joint
density.
}
