% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergence-functions.R
\name{divergence_measures}
\alias{divergence_measures}
\title{Calculate specified divergence measures for each posterior}
\usage{
divergence_measures(draws1, draws2, measure, measure_args = list(), ...)
}
\arguments{
\item{draws1}{draws of first distribution}

\item{draws2}{draws of second distribution}

\item{measure}{divergence measure}

\item{measure_args}{arguments for divergence measure}

\item{...}{unused}
}
\value{
a tibble
}
\description{
Calculate specified divergence measures for each posterior
}
