% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exptilt.R
\name{exptilt-overview}
\alias{exptilt-overview}
\alias{exptilt}
\alias{exptilt.default}
\alias{exptilt.priorsense_data}
\title{Exponential tilting perturbation}
\usage{
exptilt(x, ...)

\method{exptilt}{default}(
  x,
  tilted_variable,
  eta,
  tilt_fun = I,
  moment_match = FALSE,
  k_threshold = NULL,
  resample = FALSE,
  transform = NULL,
  prediction = NULL,
  variable = NULL,
  variables = NULL,
  tilt_fun_args = NULL,
  ...
)

\method{exptilt}{priorsense_data}(
  x,
  tilted_variable,
  eta,
  tilt_fun = I,
  moment_match = FALSE,
  k_threshold = NULL,
  resample = FALSE,
  transform = NULL,
  prediction = NULL,
  variable = NULL,
  variables = NULL,
  tilt_fun_args = NULL,
  ...
)
}
\arguments{
\item{x}{A fitted model object.}

\item{...}{Further arguments passed to internal functions.}

\item{prediction}{Function taking the model fit and returning a
draws_df of predictions to be appended to the posterior draws}

\item{variable}{Vector of variable names to return estimated
posterior draws for. If \code{NULL} all variables will be included.}

\item{variables}{Alias of \code{variable}.}

\item{prior_selection}{Vector specifying partitions of component to be
included in power-scaling. Default is NULL, which takes all
partitions. If this is a character, then it is appended to the
variable name (specified by \code{log_prior_name}) with an \verb{_}
between them. If numeric, then it is appended inside \verb{[]}.}

\item{likelihood_selection}{Vector specifying partitions of component to be
included in power-scaling. Default is NULL, which takes all
partitions. If this is a character, then it is appended to the
variable name (specified by \code{log_lik_name}) with an \verb{_}
between them. If numeric, then it is appended inside \verb{[]}.}
}
\value{
An \code{exptilted_draws} or \code{exptilted_sequence} object,
which contains the estimated posterior draws resulting from the
power-scaling perturbations and details of the perturbation and
estimation methods.
}
\description{
Estimate posterior draws based on exponential perturbations of
prior using importance sampling (and optionally moment matching).
}
\examples{
ex <- example_powerscale_model()

exptilt(ex$draws, tilted_variable = "sigma", eta = 0.5)

}
\references{
Kallioinen, N., Paananen, T., Bürkner, P-C., Vehtari, A. (2023).
Detecting and diagnosing prior and likelihood sensitivity with
power-scaling perturbations.  \emph{Statistics and
Computing}. 34(57).  \code{doi:10.1007/s11222-023-10366-5}

Vehtari, A., Simpson, D., Gelman, A., Yao, Y., and Gabry,
J. (2024).  Pareto smoothed importance sampling. \emph{Journal of
Machine Learning Research}. 25(72).
\code{https://jmlr.org/papers/v25/19-556.html}

Paananen, T., Piironen, J., Bürkner, P-C., Vehtari, A. (2021).
Implicitly adaptive importance sampling.  \emph{Statistics and
Computing}. 31(16).  \code{doi:10.1007/s11222-020-09982-2}
}
