% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_log_prior.R
\name{extract_log_prior}
\alias{extract_log_prior}
\alias{extract_log_prior.stanfit}
\alias{extract_log_prior.brmsfit}
\alias{extract_log_prior.CmdStanFit}
\title{Extract log prior from fitted model}
\usage{
extract_log_prior(x, ...)

\method{extract_log_prior}{stanfit}(x, parameter_name = "log_prior", ...)

\method{extract_log_prior}{brmsfit}(x, parameter_name = "log_prior", ...)

\method{extract_log_prior}{CmdStanFit}(x, parameter_name = "log_prior", ...)
}
\arguments{
\item{x}{Model fit.}

\item{...}{Arguments passed to individual methods.}

\item{parameter_name}{Name of parameter in Stan model
corresponding to log prior, default is "log_prior".}
}
\value{
A draws_array object containing log_prior values.
}
\description{
Extract log prior from variable in fitted Stan model.
}
