% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_alpha_threshold.R
\name{find_alpha_threshold}
\alias{find_alpha_threshold}
\title{Find the value of alpha furthest from 1 which gives the specified
pareto-k value}
\usage{
find_alpha_threshold(x, ...)
}
\arguments{
\item{fit}{a model fit of class \code{brmsfit}}

\item{variables}{variables to report}

\item{component}{string specifying which component to investigate,
\code{likelihood} or \code{prior}.}

\item{range}{vector of length 2 specifying the range of alphas to
consider, both values must be less than 1 or greater than 1}

\item{k_threshold}{highest acceptable pareto-k value}

\item{epsilon}{value to define differences}

\item{moment_match}{moment match}
}
\value{
numeric
}
\description{
Find the value of alpha furthest from 1 which gives the specified
pareto-k value
}
