% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_joint_log_lik.R
\name{joint_log_lik}
\alias{joint_log_lik}
\alias{joint_log_lik_stanfit}
\alias{joint_log_lik_brmsfit}
\alias{joint_log_lik_CmdStanFit}
\title{Extract joint log likelihood from fitted model}
\usage{
joint_log_lik_stanfit(x, parameter_name = "log_lik", ...)

joint_log_lik_brmsfit(x, parameter_name = "log_lik", ...)

joint_log_lik_CmdStanFit(x, parameter_name = "log_lik", ...)
}
\arguments{
\item{x}{Model fit.}

\item{parameter_name}{Name of parameter in Stan model
corresponding to log prior, default is "log_prior".}

\item{...}{Arguments passed to individual methods.}
}
\value{
A draws_array object containing log_prior values.
}
\description{
Extract joint log likelihood from fitted model
}
