% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergence-functions.R
\name{js_dist}
\alias{js_dist}
\title{Jensen-Shannon distance}
\usage{
js_dist(x, y, x_weights, y_weights, ...)
}
\arguments{
\item{x}{Samples from first distribution}

\item{y}{Samples from second distribution}

\item{x_weights}{Weights of first distribution}

\item{y_weights}{Weights of second distribution}

\item{...}{unused}
}
\value{
numeric
}
\description{
Jensen-Shannon distance
}
