% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_divergence_measures.R
\name{js_div}
\alias{js_div}
\title{Jensen-Shannon divergence}
\usage{
js_div(x, y, x_weights, y_weights, ...)
}
\arguments{
\item{x}{Samples from first distribution}

\item{y}{Samples from second distribution}

\item{x_weights}{Weights of first distribution}

\item{y_weights}{Weights of second distribution}

\item{...}{unused}
}
\value{
numeric
}
\description{
Jensen-Shannon divergence
}
