% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergence-functions.R
\name{kl_div}
\alias{kl_div}
\title{Kullback-Leibler divergence}
\usage{
kl_div(x, y, x_weights, y_weights, ...)
}
\arguments{
\item{x}{Samples from first distribution}

\item{y}{Samples from second distribution}

\item{x_weights}{Weights of first distribution}

\item{y_weights}{Weights of second distribution}

\item{...}{unused}
}
\value{
numeric value of approximate KL(p_x||p_y) based on
estimated densitites
}
\description{
Kullback-Leibler divergence
}
