% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergence-functions.R
\name{kl_multi_div}
\alias{kl_multi_div}
\title{Multivariate Kullback-Leibler divergence}
\usage{
kl_multi_div(weights, ...)
}
\arguments{
\item{weights}{importance weights (unnormalized)}

\item{...}{unused}
}
\description{
Multivariate Kullback-Leibler divergence
}
