% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergence-functions.R
\name{ks_dist}
\alias{ks_dist}
\title{Kolmogorov-Smirnov distance}
\usage{
ks_dist(x, y, x_weights, y_weights, ...)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{x_weights}{vector of weights}

\item{y_weights}{vector of weights}

\item{...}{ununsed}
}
\value{
numeric value of Kolmogorov Smirnov distance
}
\description{
Kolmogorov-Smirnov distance
}
