% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms-functions.R, R/extract_joint_log_lik.R
\name{log_lik_draws.brmsfit}
\alias{log_lik_draws.brmsfit}
\alias{log_lik_draws}
\alias{log_lik_draws.stanfit}
\alias{log_lik_draws.CmdStanFit}
\alias{log_lik_draws.draws}
\title{Extract log likelihood from fitted model}
\usage{
\method{log_lik_draws}{brmsfit}(x, ...)

log_lik_draws(x, ...)

\method{log_lik_draws}{stanfit}(x, log_lik_name = "log_lik", ...)

\method{log_lik_draws}{CmdStanFit}(x, log_lik_name = "log_lik", ...)

\method{log_lik_draws}{draws}(x, log_lik_name = "log_lik", ...)
}
\arguments{
\item{x}{Model fit.}

\item{...}{Arguments passed to individual methods.}

\item{log_lik_name}{Name of parameter in Stan model
corresponding to log likelihood, default is "log_lik".}
}
\value{
A draws_array object containing log_lik values.
}
\description{
Extract log likelihood from fitted model
}
