% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms-functions.R, R/extract_log_prior.R
\name{log_prior_brmsfit}
\alias{log_prior_brmsfit}
\alias{log_prior}
\alias{log_prior_stanfit}
\alias{log_prior_CmdStanFit}
\title{Extract log prior from fitted model}
\usage{
log_prior_brmsfit(x, log_prior_name = "lprior", ...)

log_prior_stanfit(x, log_prior_name = "lprior", ...)

log_prior_CmdStanFit(x, log_prior_name = "lprior", ...)
}
\arguments{
\item{x}{Model fit.}

\item{log_prior_name}{Name of parameter in Stan model
corresponding to log prior, default is "lprior".}

\item{...}{Arguments passed to individual methods.}
}
\value{
A draws_array object containing log_prior values.
}
\description{
Extract log prior from variable in fitted Stan model.
}
