% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_divergence.R
\name{measure_divergence}
\alias{measure_divergence}
\title{Calculate specified divergence measures for each posterior}
\usage{
measure_divergence(draws1, draws2, measure, measure_args = list(), ...)
}
\arguments{
\item{draws1}{draws of first distribution}

\item{draws2}{draws of second distribution}

\item{measure}{divergence measure}

\item{measure_args}{arguments for divergence measure}

\item{...}{unused}
}
\value{
a tibble
}
\description{
Calculate specified divergence measures for each posterior
}
