% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_divergence_measures.R
\name{mv_kl_div}
\alias{mv_kl_div}
\title{Multivariate Kullback-Leibler divergence}
\usage{
mv_kl_div(weights, ...)
}
\arguments{
\item{weights}{importance weights (unnormalized)}

\item{...}{unused}
}
\description{
Multivariate Kullback-Leibler divergence
}
