% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_divergence_measures.R
\name{mv_wasserstein_dist}
\alias{mv_wasserstein_dist}
\title{Multivariate Wasserstein distance}
\usage{
mv_wasserstein_dist(
  draws1,
  draws2,
  weights1 = NULL,
  weights2 = NULL,
  subsample_size = 100,
  ...
)
}
\arguments{
\item{draws1}{draws from first distribution}

\item{draws2}{draws from second distribution}

\item{weights1}{weights for first distribution}

\item{weights2}{weights for second distribution}

\item{subsample_size}{size of subsamples}

\item{...}{unused}
}
\description{
Multivariate Wasserstein distance
}
