% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_quantities.R
\name{n_eff_var}
\alias{n_eff_var}
\title{Importance sampling effective sample size diagnostic for
computing the variance of a parameter}
\usage{
n_eff_var(x, weights, ...)
}
\arguments{
\item{x}{vector of values}

\item{weights}{vector of weights}

\item{...}{unused}
}
\value{
numeric
}
\description{
Importance sampling effective sample size diagnostic for
computing the variance of a parameter
}
