% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerscale_gradients.R
\name{powerscale_gradients}
\alias{powerscale_gradients}
\title{Power-scale gradients}
\usage{
powerscale_gradients(
  fit,
  variables = NULL,
  component = c("prior", "likelihood"),
  type = c("quantities", "divergence"),
  lower_alpha = 0.99,
  upper_alpha = 1.01,
  div_measure = "cjs_dist",
  measure_args = list(),
  is_method = "psis",
  moment_match = FALSE,
  k_threshold = 0.5,
  resample = FALSE,
  transform = FALSE,
  scale = FALSE,
  log_prior_fn = calculate_log_prior,
  joint_log_lik_fn = extract_joint_log_lik,
  ...
)
}
\arguments{
\item{fit}{Model fit of class brmsfit, stanfit or CmdStanFit}

\item{variables}{Variables to compute sensitivity of. If NA
(default) sensitivity is computed for all variables.}

\item{component}{Component to power-scale (prior or likelihood).}

\item{type}{type of sensitivity to measure ("distance",
"quantity").  Multiple options can be specified at the same
time.}

\item{lower_alpha}{lower power to scale component by, should be <
1 (default is 0.9).}

\item{upper_alpha}{upper power to scale component by, should be >
1 (default is 1.1).}

\item{div_measure}{The divergence measure to use. The
following methods are implemented:
\itemize{
\item \code{"cjs_dist"}: Cumulative Jensen-Shannon distance. Default
method. See function \code{cjs_dist} for more details.
\item \code{"js_dist"}: Jensen-Shannon distance.
\item \code{"js_div"}: Jensen-Shannon divergence.
\item \code{"hellinger_dist"}: Hellinger distance.
\item \code{"kl_dist"}: Kullback-Leibler distance.
\item \code{"kl_div"}: Kullback-Leibler divergence.
\item \code{"ks_dist"}: Kolmogorov-Smirnov distance.
\item \code{"hellinger_dist"}: Hellinger distance.
\item \code{"ws_dist"}: Wassterstein distance (pass \code{measure_args = list(p = N)})
for a different order, where N is the order.
}}

\item{measure_args}{Named list of further arguments passed to divergence measure functions.}

\item{is_method}{The importance sampling method to use. The
following methods are available:
\itemize{
\item \code{"psis"}: Pareto-Smoothed Importance Sampling (PSIS). Default
method.
\item \code{"tis"}: Truncated Importance Sampling (TIS) with truncation at
\code{sqrt(S)}, where \code{S} is the number of posterior draws.
\item \code{"sis"}: Standard Importance Sampling (SIS).
}

For further details, see the \code{loo} package.}

\item{moment_match}{Logical; Indicate whether or not moment
matching should be performed. Can only be TRUE if \code{is_method} is
"psis".}

\item{k_threshold}{Threshold value for Pareto k values above which
the moment matching algorithm is used. Default is 0.5.}

\item{resample}{Logical; Indicate whether or not draws should be
resampled based on calculated importance weights.}

\item{transform}{Indicate transformation of posterior
draws. Either "scale" or "spherize".}

\item{scale}{logical scale quantity gradients by base posterior
standard deviation.}

\item{log_prior_fn}{A function that takes as input the model fit
and returns the log prior values. Provided functions are
\code{calculate_log_prior} (requires model to be fit with rstan and
in the same R session) and \code{extract_log_prior} (requires a
variable in the Stan model code to correspond to the joint log
prior, by default named "log_prior").}

\item{joint_log_lik_fn}{A function that takes as input the model
fit and returns the joint log likelihood values.}

\item{...}{Further arguments passed to functions.}
}
\value{
Maximum of the absolute derivatives above and below alpha
= 1.
}
\description{
Calculate the numerical derivative of the posterior to the
specified component (prior or likelihood) of the model fit. This
is done using importance sampling (and optionally moment matching)
to approximate the posteriors that result from power-scaling the
specified component distribution (prior or likelihood).
}
