% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerscale.R
\name{resample_draws.powerscaled_draws}
\alias{resample_draws.powerscaled_draws}
\title{Resample power-scaled draws based on weights}
\usage{
\method{resample_draws}{powerscaled_draws}(x, ...)
}
\arguments{
\item{x}{powerscaled_draws}

\item{...}{unused}
}
\description{
Resample object of class powerscaled_draws if not already
resampled
}
