% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_log_ratio.R
\name{scaled_log_ratio}
\alias{scaled_log_ratio}
\title{Calculate importance ratios based on scaling of component}
\usage{
scaled_log_ratio(
  x,
  component,
  alpha,
  log_prior_fn = calculate_log_prior,
  joint_log_lik_fn = extract_joint_log_lik,
  ...
)
}
\arguments{
\item{x}{a model fit}

\item{component}{component to scale}

\item{alpha}{scaling factor}

\item{log_prior_fn}{function to extract log_prior}

\item{joint_log_lik_fn}{function to extract joing log lik}

\item{...}{other arguments}
}
\value{
log ratio
}
\description{
Calculate importance ratios based on scaling of component
}
