% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherize.R
\name{spherize_draws}
\alias{spherize_draws}
\title{Transform draws to be spherical}
\usage{
spherize_draws(draws, spherize_method = "PCA-cor", ...)
}
\arguments{
\item{draws}{draws to be transformed}

\item{spherize_method}{transformation method}

\item{...}{unused}
}
\value{
transformed draws
}
\description{
Transform draws to be spherical
}
