% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spherize.R
\name{spherize_matrix}
\alias{spherize_matrix}
\title{Transform matrix to be spherical}
\usage{
spherize_matrix(x, spherize_method = "PCA", ...)
}
\arguments{
\item{x}{matrix to be transformed}

\item{...}{unused}

\item{method}{transformation method}
}
\value{
transformed matrix
}
\description{
Transform matrix to be spherical
}
