% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_draws.R
\name{summarise_draws.powerscaled_sequence}
\alias{summarise_draws.powerscaled_sequence}
\title{Summarise draws}
\usage{
\method{summarise_draws}{powerscaled_sequence}(
  .x,
  ...,
  .args = list(),
  div_measures = "cjs_dist",
  measure_args = list(),
  resample = FALSE
)
}
\arguments{
\item{.x}{powerscaled_sequence}

\item{...}{summary functions}

\item{.args}{additional arguments to summary functions}

\item{div_measures}{divergence measures}

\item{measure_args}{arguments for divergence measures}

\item{resample}{resample}
}
\value{
powerscaled_sequence_summary
}
\description{
Summarise draws
}
