% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_quantities.R
\name{weighted_summary_measures}
\alias{weighted_summary_measures}
\title{Weighted summary measures}
\usage{
weighted_summary_measures(x)
}
\arguments{
\item{x}{draws object to extract weights from}
}
\value{
Vector of formulas for use with \code{posterior::summarise_draws()}
}
\description{
Returns weighted versions of
\code{posterior::default_summary_measures()} to be used with
\code{posterior::summarise_draws()}.
}
\keyword{internal}
