% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerscale.R
\name{weights.powerscaled_draws}
\alias{weights.powerscaled_draws}
\title{Extract weights from power-scaled draws}
\usage{
\method{weights}{powerscaled_draws}(object, ...)
}
\arguments{
\item{object}{powerscaled_draws}

\item{...}{unused}
}
\description{
Extract weights from an object of class powerscaled_draws
}
