% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiten_draws.R
\name{whiten_draws}
\alias{whiten_draws}
\title{Transform draws to be spherical}
\usage{
whiten_draws(draws, whitening_method = "PCA-cor", ...)
}
\arguments{
\item{draws}{draws to be transformed}

\item{whitening_method}{transformation method}

\item{...}{unused}
}
\value{
transformed draws
}
\description{
Transform draws to be spherical
}
