% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiten_draws.R
\name{whiten_matrix}
\alias{whiten_matrix}
\title{Transform matrix to be spherical}
\usage{
whiten_matrix(x, whitening_method = "PCA-cor", ...)
}
\arguments{
\item{x}{matrix to be transformed}

\item{whitening_method}{transformation method}

\item{...}{unused}
}
\value{
transformed matrix
}
\description{
Transform matrix to be spherical
}
