% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_divergence_measures.R
\name{ws_dist}
\alias{ws_dist}
\title{Wasserstein distance}
\usage{
ws_dist(x, y, x_weights, y_weights, p = 1, ...)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{x_weights}{vector of weights}

\item{y_weights}{vector of weights}

\item{p}{degree}

\item{...}{unused}
}
\value{
numeric value of Wassterstein distance
}
\description{
Wasserstein distance
}
