% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms-functions.R
\name{create-powerscaling-data}
\alias{create-powerscaling-data}
\alias{create_priorsense_data.brmsfit}
\title{Create powerscaling data}
\usage{
\method{create_priorsense_data}{brmsfit}(x, ...)
}
\description{
Create powerscaling data
}
