% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_powerscaling_data.R
\name{create-powerscaling-data}
\alias{create-powerscaling-data}
\alias{create_powerscaling_data}
\alias{create_powerscaling_data.default}
\alias{create_powerscaling_data.stanfit}
\alias{create_powerscaling_data.CmdStanFit}
\title{Create power-scaling data structure}
\usage{
create_powerscaling_data(x, ...)

\method{create_powerscaling_data}{default}(
  x,
  get_draws,
  log_prior,
  log_lik,
  constrain_pars,
  unconstrain_pars,
  log_prob_upars,
  log_ratio_upars,
  prediction = NULL,
  ...
)

\method{create_powerscaling_data}{stanfit}(x, ...)

\method{create_powerscaling_data}{CmdStanFit}(x, ...)
}
\arguments{
\item{x}{a fit object}

\item{...}{arguments passed to methods}

\item{get_draws}{function to extract posterior draws from fit}

\item{log_prior}{draws from log prior}

\item{log_lik}{draws from log likelihood}

\item{constrain_pars}{function that transforms unconstrained
parameters to constrained space}

\item{unconstrain_pars}{function that transforms constrained
parameters to unconstrained space}

\item{log_prob_upars}{function that returns log density given
unconstrained parameter values}

\item{log_ratio_upars}{function that returns log density ratio for
importance sampling given unconstrained parameter values}

\item{prediction}{function that returns predictions from x to be
added to the draws}
}
\value{
A \code{powerscaling_data} object, which contains the data and
functions to run power-scaling sensitivity analyses.
}
\description{
Create a data structure that contains all required data and
functions for powerscaling
}
