% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_joint_log_lik.R
\name{log_lik_draws}
\alias{log_lik_draws}
\title{Extract log likelihood from fitted model}
\usage{
log_lik_draws(x, ...)
}
\arguments{
\item{x}{Model fit.}

\item{...}{Arguments passed to individual methods.}

\item{log_lik_name}{Name of parameter in Stan model
corresponding to log likelihood, default is "log_lik".}
}
\value{
A draws_array object containing log_prior values.
}
\description{
Extract log likelihood from fitted model
}
