% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_log_prior.R
\name{log_prior_draws}
\alias{log_prior_draws}
\title{Extract log prior from fitted model}
\usage{
log_prior_draws(x, ...)
}
\arguments{
\item{x}{Model fit.}

\item{...}{Arguments passed to individual methods.}

\item{log_prior_name}{Name of parameter in Stan model
corresponding to log prior, default is "lprior".}
}
\value{
A draws_array object containing log_prior values.
}
\description{
Extract log prior from variable in fitted Stan model.
}
