% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms-functions.R, R/extract_log_prior.R
\name{log_prior_draws.brmsfit}
\alias{log_prior_draws.brmsfit}
\alias{log_prior_draws}
\alias{log_prior_draws.stanfit}
\alias{log_prior_draws.CmdStanFit}
\alias{log_prior_draws.draws}
\title{Extract log prior from fitted model}
\usage{
\method{log_prior_draws}{brmsfit}(x, log_prior_name = "lprior", ...)

log_prior_draws(x, ...)

\method{log_prior_draws}{stanfit}(x, log_prior_name = "lprior", ...)

\method{log_prior_draws}{CmdStanFit}(x, log_prior_name = "lprior", ...)

\method{log_prior_draws}{draws}(x, log_prior_name = "lprior", ...)
}
\arguments{
\item{x}{Model fit.}

\item{log_prior_name}{Name of parameter in Stan model
corresponding to log prior, default is "lprior".}

\item{...}{Arguments passed to individual methods.}
}
\value{
A draws_array obejct containing log_prior values.
}
\description{
Extract log prior from variable in fitted Stan model.
}
