% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_diagnostics.R
\name{n_eff_mean}
\alias{n_eff_mean}
\title{Importance sampling effective sample size diagnostic for
computing the mean of a parameter}
\usage{
n_eff_mean(x, log_ratios, ...)
}
\arguments{
\item{x}{vector of values}

\item{log_ratios}{vector of standard importance sampling weights}

\item{...}{unused}
}
\value{
numeric
}
\description{
Importance sampling effective sample size diagnostic for
computing the mean of a parameter
}
