% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_diagnostics.R
\name{pareto_k_mean}
\alias{pareto_k_mean}
\title{Pareto-k diagnostic for computing the mean of a parameter}
\usage{
pareto_k_mean(x, log_ratios, ...)
}
\arguments{
\item{x}{vector of values}

\item{log_ratios}{vector of standard importance sampling weights}

\item{...}{unused}
}
\value{
numeric
}
\description{
Pareto-k diagnostic for computing the mean of a parameter
}
