% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_quantities.R
\name{pareto_k_var}
\alias{pareto_k_var}
\title{Pareto-k diagnostic for computing the variance of a parameter}
\usage{
pareto_k_var(x, weights, ...)
}
\arguments{
\item{x}{vector of values}

\item{weights}{vector of weights}

\item{...}{unused}
}
\value{
numeric
}
\description{
Pareto-k diagnostic for computing the variance of a parameter
}
