% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerscale.R, R/powerscale_sequence.R
\name{powerscale-overview}
\alias{powerscale-overview}
\alias{powerscale}
\alias{powerscale_sequence}
\title{Prior/likelihood power-scaling perturbation}
\usage{
powerscale(
  fit,
  alpha,
  variables = NA,
  component = "prior",
  is_method = "psis",
  moment_match = FALSE,
  k_threshold = 0.5,
  resample = FALSE,
  transform = FALSE,
  log_prior_fn = calculate_log_prior,
  joint_log_lik_fn = extract_joint_log_lik,
  ...
)

powerscale_sequence(
  fit,
  lower_alpha = 0.5,
  upper_alpha = 1/lower_alpha,
  alpha_step = 0.1,
  variables = NA,
  component = c("prior", "likelihood"),
  is_method = "psis",
  moment_match = FALSE,
  k_threshold = 0.5,
  resample = FALSE,
  log_prior_fn = calculate_log_prior,
  joint_log_lik_fn = extract_joint_log_lik,
  transform = FALSE,
  ...
)
}
\arguments{
\item{fit}{A fitted model object of class \code{stanfit}, \code{brmsfit} or
\code{CmdStanFit}.}

\item{alpha}{Value by which to power-scale specified component
(likelihood/prior).}

\item{variables}{Vector of variable names to return estimated
posterior draws for.}

\item{component}{Component to be power-scaled (either "prior" or
"likelihood"). For powerscale_sequence, this can be both "prior"
and "likelihood".}

\item{is_method}{The importance sampling method to use. The
following methods are available:
\itemize{
\item \code{"psis"}: Pareto-Smoothed Importance Sampling (PSIS). Default
method.
\item \code{"tis"}: Truncated Importance Sampling (TIS) with truncation at
\code{sqrt(S)}, where \code{S} is the number of posterior draws.
\item \code{"sis"}: Standard Importance Sampling (SIS).
}

For further details, see the \code{loo} package.}

\item{moment_match}{Logical; Indicate whether or not moment
matching should be performed. Can only be TRUE if \code{is_method} is
"psis".}

\item{k_threshold}{Threshold value for Pareto k values above which
the moment matching algorithm is used. Default is 0.5.}

\item{resample}{Logical; Indicate whether or not draws should be
resampled based on calculated importance weights.}

\item{transform}{Indicate transformation of posterior
draws. Either "scale" or "spherize".}

\item{log_prior_fn}{A function that takes as input the model fit
and returns the log prior values. Provided functions are
\code{calculate_log_prior} (requires model to be fit with rstan and
in the same R session) and \code{extract_log_prior} (requires a
variable in the Stan model code to correspond to the joint log
prior, by default named "log_prior").}

\item{joint_log_lik_fn}{A function that takes as input the model
fit and returns the joint log likelihood values.}

\item{...}{Further arguments passed to the custom functions
documented above.}

\item{lower_alpha}{Lower power-scaling alpha value in sequence.}

\item{upper_alpha}{Upper power-scaling alpha value in sequence.}

\item{alpha_step}{Step size of power-scaling alphas in sequence.}
}
\value{
A \code{powerscaled_draws} or \code{powerscaled_sequence} object,
which contains the estimated posterior draws resulting from the
power-scaling perturbations and details of the perturbation and
estimation methods.
}
\description{
Estimate posterior draws based on power-scaling perturbations of
prior or likelihood using importance sampling (and optionally moment matching).
}
\references{
Kallioinen, N., Paananen, T., Bürkner, P-C., Vehtari, A. (in
prep).  Detecting and diagnosing prior and likelihood sensitivity
with power-scaling perturbations.

Vehtari, A., Simpson, D., Gelman, A., Yao, Y., and Gabry,
J. (2019).  Pareto smoothed importance sampling.  \href{https://arxiv.org/abs/1507}{preprint arXiv:1507.02646}

Paananen, T., Piironen, J., Bürkner, P-C., Vehtari, A. (2021).
Implicitly adaptive importance sampling.  \emph{Statistics and
Computing}. 31(16).  \code{doi:10.1007/s11222-020-09982-2}
}
