% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms-functions.R, R/powerscale.R,
%   R/powerscale_sequence.R
\name{powerscale.brmsfit}
\alias{powerscale.brmsfit}
\alias{powerscale_sequence.brmsfit}
\alias{powerscale-overview}
\alias{powerscale}
\alias{powerscale.powerscaling_data}
\alias{powerscale.CmdStanFit}
\alias{powerscale.stanfit}
\alias{powerscale_sequence}
\alias{powerscale_sequence.CmdStanFit}
\alias{powerscale_sequence.stanfit}
\alias{powerscale_sequence.powerscaling_data}
\title{Prior/likelihood power-scaling perturbation}
\usage{
\method{powerscale}{brmsfit}(x, component, alpha, ...)

\method{powerscale_sequence}{brmsfit}(x, ...)

powerscale(x, ...)

\method{powerscale}{powerscaling_data}(
  x,
  alpha,
  component,
  is_method = "psis",
  moment_match = FALSE,
  k_threshold = 0.5,
  resample = FALSE,
  transform = FALSE,
  prediction = NULL,
  variable = NULL,
  ...
)

\method{powerscale}{CmdStanFit}(x, component, alpha, ...)

\method{powerscale}{stanfit}(x, component, alpha, ...)

powerscale_sequence(x, ...)

\method{powerscale_sequence}{CmdStanFit}(x, ...)

\method{powerscale_sequence}{stanfit}(x, ...)

\method{powerscale_sequence}{powerscaling_data}(
  x,
  lower_alpha = 0.8,
  upper_alpha = 1/lower_alpha,
  length = 9,
  variable = NULL,
  component = c("prior", "likelihood"),
  is_method = "psis",
  moment_match = FALSE,
  k_threshold = 0.5,
  resample = FALSE,
  transform = FALSE,
  auto_alpha_range = FALSE,
  symmetric = TRUE,
  ...
)
}
\arguments{
\item{x}{A fitted model object.}

\item{component}{Component to be power-scaled (either "prior" or
"likelihood"). For powerscale_sequence, this can be both "prior"
and "likelihood".}

\item{alpha}{Value by which to power-scale specified
component. (likelihood/prior).}

\item{...}{Further arguments passed to the custom functions
documented above.}

\item{is_method}{The importance sampling method to use. The
following methods are available:
\itemize{
\item \code{"psis"}: Pareto-Smoothed Importance Sampling (PSIS). Default
method.
\item \code{"tis"}: Truncated Importance Sampling (TIS) with truncation at
\code{sqrt(S)}, where \code{S} is the number of posterior draws.
\item \code{"sis"}: Standard Importance Sampling (SIS).
}

For further details, see the \code{loo} package.}

\item{moment_match}{Logical; Indicate whether or not moment
matching should be performed. Can only be TRUE if \code{is_method} is
"psis".}

\item{k_threshold}{Threshold value for Pareto k values above which
the moment matching algorithm is used. Default is 0.5.}

\item{resample}{Logical; Indicate whether or not draws should be
resampled based on calculated importance weights.}

\item{transform}{Indicate a transformation of posterior
draws to perform before sensitivity analysis. Either "scale" or "whiten".}

\item{prediction}{Function taking the model fit and returning a
draws_df of predictions to be appended to the posterior draws}

\item{variable}{Vector of variable names to return estimated
posterior draws for.}

\item{lower_alpha}{Lower power-scaling alpha value in sequence.}

\item{upper_alpha}{Upper power-scaling alpha value in sequence.}

\item{length}{Length of alpha sequence.}

\item{auto_alpha_range}{Boolean. Restrict range to ensure Pareto-k
values below threshold?}

\item{symmetric}{Boolean. Should the alpha range be symmetrical
around alpha = 1, on log-space?}
}
\value{
A \code{powerscaled_draws} or \code{powerscaled_sequence} object,
which contains the estimated posterior draws resulting from the
power-scaling perturbations and details of the perturbation and
estimation methods.
}
\description{
Estimate posterior draws based on power-scaling perturbations of
prior or likelihood using importance sampling (and optionally moment matching).
}
\references{
Kallioinen, N., Paananen, T., Bürkner, P-C., Vehtari, A. (in
prep).  Detecting and diagnosing prior and likelihood sensitivity
with power-scaling perturbations.

Vehtari, A., Simpson, D., Gelman, A., Yao, Y., and Gabry,
J. (2019).  Pareto smoothed importance sampling.  \href{https://arxiv.org/abs/1507}{preprint arXiv:1507.02646}

Paananen, T., Piironen, J., Bürkner, P-C., Vehtari, A. (2021).
Implicitly adaptive importance sampling.  \emph{Statistics and
Computing}. 31(16).  \code{doi:10.1007/s11222-020-09982-2}
}
