% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerscale_sensitivity.R
\name{powerscale-sensitivity}
\alias{powerscale-sensitivity}
\alias{powerscale_sensitivity}
\alias{powerscale_sensitivity.default}
\alias{powerscale_sensitivity.powerscaling_data}
\alias{powerscale_sensitivity.CmdStanFit}
\alias{powerscale_sensitivity.stanfit}
\alias{powerscale_sensitivity.brmsfit}
\title{Power-scaling sensitivity analysis}
\usage{
powerscale_sensitivity(x, ...)

\method{powerscale_sensitivity}{default}(x, ...)

\method{powerscale_sensitivity}{powerscaling_data}(
  x,
  variable = NULL,
  lower_alpha = 0.99,
  upper_alpha = 1.01,
  div_measure = "cjs_dist",
  measure_args = list(),
  component = c("prior", "likelihood"),
  sensitivity_threshold = 0.05,
  is_method = "psis",
  moment_match = FALSE,
  k_threshold = 0.5,
  resample = FALSE,
  transform = FALSE,
  prediction = NULL,
  ...
)

\method{powerscale_sensitivity}{CmdStanFit}(x, ...)

\method{powerscale_sensitivity}{stanfit}(x, ...)

\method{powerscale_sensitivity}{brmsfit}(x, ...)
}
\arguments{
\item{x}{Model fit object or powerscaling_data object.}

\item{...}{Further arguments passed to functions.}

\item{variable}{Character vector of variables to check.}

\item{lower_alpha}{Lower alpha value for gradient calculation.}

\item{upper_alpha}{Upper alpha value for gradient calculation.}

\item{div_measure}{The divergence measure to use. The
following methods are implemented:
\itemize{
\item \code{"cjs_dist"}: Cumulative Jensen-Shannon distance. Default
method. See function \code{cjs_dist} for more details.
\item \code{"js_dist"}: Jensen-Shannon distance.
\item \code{"js_div"}: Jensen-Shannon divergence.
\item \code{"hellinger_dist"}: Hellinger distance.
\item \code{"kl_dist"}: Kullback-Leibler distance.
\item \code{"kl_div"}: Kullback-Leibler divergence.
\item \code{"ks_dist"}: Kolmogorov-Smirnov distance.
\item \code{"hellinger_dist"}: Hellinger distance.
\item \code{"ws_dist"}: Wassterstein distance (pass \code{measure_args = list(p = N)})
for a different order, where N is the order.
}}

\item{measure_args}{Named list of further arguments passed to divergence measure functions.}

\item{component}{Character vector specifying component(s) to scale (default is both "prior" and
"likelihood").}

\item{sensitivity_threshold}{Threshold for flagging variable as
sensitive to power-scaling.}

\item{is_method}{The importance sampling method to use. The
following methods are available:
\itemize{
\item \code{"psis"}: Pareto-Smoothed Importance Sampling (PSIS). Default
method.
\item \code{"tis"}: Truncated Importance Sampling (TIS) with truncation at
\code{sqrt(S)}, where \code{S} is the number of posterior draws.
\item \code{"sis"}: Standard Importance Sampling (SIS).
}

For further details, see the \code{loo} package.}

\item{moment_match}{Logical; Indicate whether or not moment
matching should be performed. Can only be TRUE if \code{is_method} is
"psis".}

\item{k_threshold}{Threshold value for Pareto k values above which
the moment matching algorithm is used. Default is 0.5.}

\item{resample}{Logical; Indicate whether or not draws should be
resampled based on calculated importance weights.}

\item{transform}{Indicate transformation of posterior
draws. Either "scale" or "spherize".}

\item{prediction}{Function taking the model fit and returning a
draws_df of predictions to be appended to the posterior draws}
}
\value{
Table of sensitivity values for each specified variable.
}
\description{
Prior/likelihood sensitivity analysis based on power-scaling perturbations.
}
\references{
Kallioinen, N., Paananen, T., Bürkner, P-C., Vehtari, A. (in
prep).  Detecting and diagnosing prior and likelihood sensitivity
with power-scaling perturbations.

Vehtari, A., Simpson, D., Gelman, A., Yao, Y., and Gabry,
J. (2019).  Pareto smoothed importance sampling.  \href{https://arxiv.org/abs/1507}{preprint arXiv:1507.02646}

Paananen, T., Piironen, J., Bürkner, P-C., Vehtari, A. (2021).
Implicitly adaptive importance sampling.  \emph{Statistics and
Computing}. 31(16).  \code{doi:10.1007/s11222-020-09982-2}
}
