% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerscale_derivative.R
\name{powerscale_derivative}
\alias{powerscale_derivative}
\title{Derivative with respect to power-scaling}
\usage{
powerscale_derivative(x, log_component, quantity = "mean", ...)
}
\arguments{
\item{x}{Posterior draws.}

\item{log_component}{Log likelihood or log prior values.}

\item{quantity}{Character specifying quantity of interest (default
is "mean"). Options are "mean", "sd", "var".}

\item{...}{unused}
}
\value{
Derivative of the quantity with respect to log of the
power-scaling factor (alpha).
}
\description{
Calculate the derivative of a quantity with respect to
power-scaling prior or likelihood.
}
\examples{
\dontrun{
example_model <- example_powerscale_model()

fit <- rstan::stan(model_code = example_model$model_code, data = example_model$data)

draws <- posterior::subset_draws(posterior::as_draws_df(fit), variable = c("mu", "sigma"))

log_prior <- as.numeric(calculate_log_prior(fit))

posterior::summarise_draws(draws, mean, ~powerscale_derivative(.x, log_prior, quantity = "mean"))
}
}
