% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative-functions.R,
%   R/powerscale_derivative.R
\name{powerscale_derivative}
\alias{powerscale_derivative}
\title{Derivative with respect to power-scaling}
\usage{
powerscale_derivative(x, log_component, quantity = "mean")

powerscale_derivative(x, log_component, quantity = "mean")
}
\arguments{
\item{x}{Vector of draws.}

\item{log_component}{Vector of log likelihood or log prior values.}

\item{quantity}{Character specifying quantity of interest. "mean",
"sd" and "var" are currently implemented.}
}
\value{
Derivative of the quantity with respect to log of the
power-scaling factor (alpha).

Derivative of the quantity with respect to the
power-scaling factor.
}
\description{
Calculate the derivative of a quantity with respect to
power-scaling prior or likelihood.

Calculate the derivative of a posterior quantity with respect to
power-scaling the prior or likelihood.
}
\examples{
example_model <- example_powerscale_model()

fit <- rstan::stan(model_code = example_model$model_code, data = example_model$data)

draws <- posterior::subset_draws(posterior::as_draws_df(fit), variable = c("mu", "sigma"))

log_prior <- as.numeric(calculate_log_prior(fit))

posterior::summarise_draws(draws, mean, ~powerscale_derivative(.x, log_prior, quantity = "mean"))

}
\references{
Kallioinen, N., Paananen, T., Bürkner, P-C., Vehtari, A. (in
prep).  Detecting and diagnosing prior and likelihood sensitivity
with power-scaling perturbations.

Vehtari, A., Simpson, D., Gelman, A., Yao, Y., and Gabry,
J. (2019).  Pareto smoothed importance sampling.  \href{https://arxiv.org/abs/1507}{preprint arXiv:1507.02646}

Paananen, T., Piironen, J., Bürkner, P-C., Vehtari, A. (2021).
Implicitly adaptive importance sampling.  \emph{Statistics and
Computing}. 31(16).  \code{doi:10.1007/s11222-020-09982-2}
}
