% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{powerscale_plots}
\alias{powerscale_plots}
\alias{powerscale_plot_dens}
\alias{powerscale_plot_ecdf}
\alias{powerscale_plot_ecdf.powerscaled_sequence}
\alias{powerscale_plot_quantities}
\alias{powerscale_plot_quantities.powerscaled_sequence}
\title{Diagnostic plots for power-scaling sensitivity}
\usage{
powerscale_plot_dens(x, ...)

powerscale_plot_ecdf(x, ...)

\method{powerscale_plot_ecdf}{powerscaled_sequence}(
  x,
  variable = NULL,
  resample = FALSE,
  auto_title = TRUE,
  ...
)

powerscale_plot_quantities(x, ...)

\method{powerscale_plot_quantities}{powerscaled_sequence}(
  x,
  variable = NULL,
  quantities = c("mean", "sd"),
  div_measure = "cjs_dist",
  resample = FALSE,
  measure_args = NULL,
  mcse = TRUE,
  quantity_args = NULL,
  auto_title = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{powerscaled_sequence} or an object
for which \code{powerscale_sequence} will first be run on.}

\item{...}{Arguments passed to \code{powerscale_sequence} if \code{x} is not
of class \code{powerscaled_sequence}.}

\item{variable}{A character vector of variable names. If \code{NULL}
(the default) all variables will be plotted.}

\item{resample}{Logical; Indicate whether or not draws should be
resampled based on calculated importance weights.}

\item{auto_title}{Logical indicating whether automatic title with
description should be included in the plot. Default is TRUE.}

\item{quantities}{A character vector specifying quantities to
plot. Options are "mean", "median", "sd", "mad", "quantile".}

\item{div_measure}{The divergence measure to use. The
following methods are implemented:
\itemize{
\item \code{"cjs_dist"}: Cumulative Jensen-Shannon distance. Default
method. See function \code{cjs_dist} for more details.
\item \code{"js_dist"}: Jensen-Shannon distance.
\item \code{"js_div"}: Jensen-Shannon divergence.
\item \code{"hellinger_dist"}: Hellinger distance.
\item \code{"kl_dist"}: Kullback-Leibler distance.
\item \code{"kl_div"}: Kullback-Leibler divergence.
\item \code{"ks_dist"}: Kolmogorov-Smirnov distance.
\item \code{"hellinger_dist"}: Hellinger distance.
\item \code{"ws_dist"}: Wassterstein distance (pass \code{measure_args = list(p = N)})
for a different order, where N is the order.
}}

\item{measure_args}{Named list of further arguments passed to divergence measure functions.}

\item{mcse}{Boolean; If TRUE will plot +/- 2 * Monte Carlo
standard error of the base quantity on the quantities plot.}

\item{quantity_args}{Named list of further arguments passed to
quantity functions. Passed as \code{.args} to
\verb{[posterior::summarise_draws]}.}
}
\value{
A ggplot object that can be further customized using the \strong{ggplot2} package.
}
\description{
Various diagnostic plots for power-scaling sensitivity. See \strong{Plot Descriptions} below for details.
}
\section{Plot Descriptions}{
 \describe{
\item{\code{powerscale_plot_dens()}}{ Kernel density plot of
power-scaled posterior draws with respect to power-scaling.  }
\item{\code{powerscale_plot_ecdf()}}{ Empirical cumulative
distribution function plot of power-scaled posterior draws with
respect to power-scaling.  }
\item{\code{powerscale_plot_quantities()}}{ Plot of posterior
quantities with respect to power-scaling.  } }
}

