% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorsense-package.R
\docType{package}
\name{priorsense-package}
\alias{priorsense-package}
\alias{priorsense}
\title{priorsense: Prior Diagnostics and Sensitivity Analysis}
\description{
priorsense provides functions for prior and likelihood
sensitivity analysis in Bayesian models. Currently it implements
methods to determine the sensitivity of the posterior to power-scaling
perturbations of the prior and likelihood.
}
