% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaled_log_ratio.R
\name{scaled_log_ratio}
\alias{scaled_log_ratio}
\title{Calculate importance ratios based on scaling of component}
\usage{
scaled_log_ratio(component_draws, alpha, ...)
}
\arguments{
\item{component_draws}{draws from component to powerscale}

\item{alpha}{scaling factor}

\item{...}{unused}
}
\value{
log ratio
}
\description{
Calculate importance ratios based on scaling of component
}
\keyword{internal}
