% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_draws.R
\name{summarise_draws.powerscaled_draws}
\alias{summarise_draws.powerscaled_draws}
\title{Summarise draws}
\usage{
\method{summarise_draws}{powerscaled_draws}(
  .x,
  ...,
  .args = list(),
  base_draws = NULL,
  div_measures = "cjs_dist",
  measure_args = list(),
  resample = FALSE
)
}
\arguments{
\item{.x}{An object of class powerscaled_draws}

\item{...}{summary functions}

\item{.args}{arguments for summary functions}

\item{base_draws}{base draws}

\item{div_measures}{divergence measures}

\item{measure_args}{arguments for divergence measures}

\item{resample}{resample draws}
}
\description{
Summarise power-scaled draws
}
