% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_draws.R
\name{summarise_draws.powerscaled_sequence}
\alias{summarise_draws.powerscaled_sequence}
\title{Summarise draws}
\usage{
\method{summarise_draws}{powerscaled_sequence}(
  .x,
  ...,
  .args = list(),
  .num_args = getOption("posterior.num_args", list()),
  div_measures = "cjs_dist",
  measure_args = list(),
  resample = FALSE
)
}
\arguments{
\item{.x}{powerscaled_sequence}

\item{...}{summary functions}

\item{.args}{additional arguments to summary functions}

\item{.num_args}{(named list) Optional arguments passed to
\link[tibble:num]{num()} for pretty printing of summaries. Can be
controlled globally via the \code{posterior.num_args}
\link[base:options]{option}.}

\item{div_measures}{divergence measures}

\item{measure_args}{arguments for divergence measures}

\item{resample}{resample}
}
\value{
powerscaled_sequence_summary
}
\description{
Summarise draws
}
