% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_quantities.R
\name{weighted_quantities}
\alias{weighted_quantities}
\title{Weighted quantities}
\arguments{
\item{x}{Numeric vector to calculate quantity from.}

\item{weights}{Vector of weights corresponding to values in x.}

\item{probs}{Vector of probabilities for quantiles.}

\item{type}{Character vector specifying type of quantiles (either
"7" for Type 7 (default) or "hd" for Harrell-Davis)}

\item{...}{Currently unused.}
}
\value{
Named vector of calculated quantity.
}
\description{
Weighted version of common quantities of interest.
}
\keyword{internal}
